moar dithering; fix our approx multiply unit and get rid of *s
this should save a lot of gate area and it looks better, too. had to do
a silily hack to prevent holes in the raymarching but whatever works.

i tried to get a 3-cordic iteration version and it looks *much* better
but it doesn't meet timing and it doesn't fit :(